install redis for caching, create album cache
